requirements.txt
Django==4.1
djangorestframework==3.14.0
django-extensions==3.2.1
